+++ hoot-js/src/main/cpp/hoot/js/conflate/merging/ElementMergerJs.cpp
@@ -45,8 +45,8 @@
 #include <hoot/js/JsRegistrar.h>
 #include <hoot/js/OsmMapJs.h>
 #include <hoot/js/SystemNodeJs.h>
-#include <hoot/js/conflate/AreaMergerJs.h>
-#include <hoot/js/conflate/PoiMergerJs.h>
+#include <hoot/js/conflate/merging/AreaMergerJs.h>
+#include <hoot/js/conflate/merging/PoiMergerJs.h>
 #include <hoot/js/util/HootExceptionJs.h>
 
 // Qt
