+++ hoot-js/src/main/cpp/hoot/js/conflate/merging/MergerFactoryJs.h
