+++ hoot-js/src/main/cpp/hoot/js/conflate/merging/PoiMergerJs.h
