+++ hoot-js/src/main/cpp/hoot/js/conflate/merging/ScriptMergerCreator.cpp
@@ -29,8 +29,8 @@
 // hoot
 #include <hoot/core/util/Factory.h>
 #include <hoot/core/conflate/merging/MarkForReviewMerger.h>
-#include <hoot/js/conflate/js/ScriptMatch.h>
-#include <hoot/js/conflate/js/ScriptMerger.h>
+#include <hoot/js/conflate/matching/ScriptMatch.h>
+#include <hoot/js/conflate/merging/ScriptMerger.h>
 
 using namespace std;
 using namespace v8;
