+++ hoot-js/src/test/cpp/hoot/js/conflate/matching/ScriptMatchTest.cpp
@@ -32,8 +32,8 @@
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/MapProjector.h>
 #include <hoot/js/HootJsStable.h>
-#include <hoot/js/conflate/js/ScriptMatch.h>
-#include <hoot/js/conflate/js/ScriptMatchCreator.h>
+#include <hoot/js/conflate/matching/ScriptMatch.h>
+#include <hoot/js/conflate/matching/ScriptMatchCreator.h>
 
 // Qt
 #include <QVariant>
@@ -41,53 +41,58 @@
 #include <iostream>
 #include <string>
 
+using namespace std;
 using namespace v8;
 
 namespace hoot
 {
 
-class ScriptMatchCreatorTest : public HootTestFixture
+class ScriptMatchTest : public HootTestFixture
 {
-  CPPUNIT_TEST_SUITE(ScriptMatchCreatorTest);
-  CPPUNIT_TEST(runIsCandidateTest);
+  CPPUNIT_TEST_SUITE(ScriptMatchTest);
+  //TODO: should this test be re-enabled?
+  //CPPUNIT_TEST(cacheTest);
+  CPPUNIT_TEST(conflictTest);
   CPPUNIT_TEST_SUITE_END();
 
 public:
 
-  void runIsCandidateTest()
+  void conflictTest()
   {
-    ScriptMatchCreator uut;
-    uut.setArguments(QStringList() << "LineStringGenericTest.js");
-
     ConfigOptions co;
     conf().set(co.getUuidHelperRepeatableKey(), true);
     conf().set(co.getReaderUseFileStatusKey(), true);
     OsmMapPtr map(new OsmMap());
-    OsmMapReaderFactory::getInstance().read(
-      map, "test-files/algorithms/js/ScriptMatchTest.osm", true);
+    OsmMapReaderFactory::getInstance().read(map, "test-files/algorithms/js/ScriptMatchTest.osm",
+      true);
     MapProjector::projectToPlanar(map);
 
-    //to be a candidate: needs to be determined a candidate by the JS rules file...which for this
-    //particular rules file, the element must be linear
-    NodePtr node1 = TestUtils::createNode(map, Status::Unknown1, 10.0, 10.0, 250.0);
-    CPPUNIT_ASSERT(!uut.isMatchCandidate(node1, map));
+    // create the test scenario in ScriptMatchTest
+    // call ScriptMatch is consistent repeatedly.
 
-    NodePtr node2 = TestUtils::createNode(map, Status::Unknown1, 10.0, 10.0, 250.0);
-    QList<NodePtr> wayNodes;
-    wayNodes.append(node1);
-    wayNodes.append(node2);
-    WayPtr way1 = TestUtils::createWay(map, wayNodes, Status::Unknown1);
-    way1->getTags().appendValue("highway", "road");
-    CPPUNIT_ASSERT(uut.isMatchCandidate(way1, map));
+    boost::shared_ptr<const MatchThreshold> mt(new MatchThreshold(0.6, 0.6, 0.6));
 
-    way1 = TestUtils::createWay(map, wayNodes, Status::Unknown1);
-    way1->setStatus(Status::Unknown1);
-    way1->getTags().clear();
-    HOOT_STR_EQUALS(false, uut.isMatchCandidate(way1, map));
+    ScriptMatchCreator smc;
+    smc.setArguments(QStringList() << "LineStringGenericTest.js");
+    vector<const Match*> matches;
+    smc.createMatches(map, matches, mt);
+    HOOT_STR_EQUALS(2, matches.size());
+    HOOT_STR_EQUALS(false, matches[0]->isConflicting(*matches[1], map));
+  }
+
+  void cacheTest()
+  {
+    ScriptMatch::ConflictKey key1(ElementId::way(-2), ElementId::way(-7));
+    QHash<ScriptMatch::ConflictKey, bool> conflicts;
+
+    ScriptMatch::ConflictKey key2(ElementId::way(-2), ElementId::way(-7));
+    conflicts[key1] = true;
+
+    HOOT_STR_EQUALS(false, (bool)(conflicts.find(key2) == conflicts.end()));
   }
 
 };
 
-CPPUNIT_TEST_SUITE_NAMED_REGISTRATION(ScriptMatchCreatorTest, "quick");
+CPPUNIT_TEST_SUITE_NAMED_REGISTRATION(ScriptMatchTest, "quick");
 
 }
